backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
slowapi==0.1.9
openai==1.54.0